+++ codegen/cms/source_code/apis/ExtractApi.ts
@@ -8,6 +8,9 @@ import { isCodeInRange} from '../util';
 import {SecurityAuthentication} from '../auth/auth';
 
 
+import { ActionResponse } from '../models/ActionResponse';
+import { FileExtractRequest } from '../models/FileExtractRequest';
+import { TaskLocator } from '../models/TaskLocator';
 
 /**
  * no description
@@ -15,25 +18,73 @@ import {SecurityAuthentication} from '../auth/auth';
 export class ExtractApiRequestFactory extends BaseAPIRequestFactory {
 
     /**
-     * Extracts a zip file in the file system. The zip file will be extracted in-place and not be deleted automatically.
-     * Extracts a zip file
-     * @param path The file system location of the zip file.
+     * Extract a zip file in the developer file system. Extraction status can be checked with the `/extract/async/tasks/taskId/status` endpoint below.
+     * Extract a zip file
+     * @param fileExtractRequest 
      */
-    public async extractByPath(path: string, _options?: Configuration): Promise<RequestContext> {
+    public async doAsync(fileExtractRequest: FileExtractRequest, _options?: Configuration): Promise<RequestContext> {
         let _config = _options || this.configuration;
 
-        // verify required parameter 'path' is not null or undefined
-        if (path === null || path === undefined) {
-            throw new RequiredError("ExtractApi", "extractByPath", "path");
+        // verify required parameter 'fileExtractRequest' is not null or undefined
+        if (fileExtractRequest === null || fileExtractRequest === undefined) {
+            throw new RequiredError("ExtractApi", "doAsync", "fileExtractRequest");
         }
 
 
         // Path Params
-        const localVarPath = '/cms/v3/source-code/extract/{path}'
-            .replace('{' + 'path' + '}', encodeURIComponent(String(path)));
+        const localVarPath = '/cms/v3/source-code/extract/async';
 
         // Make Request Context
         const requestContext = _config.baseServer.makeRequestContext(localVarPath, HttpMethod.POST);
+        requestContext.setHeaderParam("Accept", "application/json, */*;q=0.8")
+
+
+        // Body Params
+        const contentType = ObjectSerializer.getPreferredMediaType([
+            "application/json"
+        ]);
+        requestContext.setHeaderParam("Content-Type", contentType);
+        const serializedBody = ObjectSerializer.stringify(
+            ObjectSerializer.serialize(fileExtractRequest, "FileExtractRequest", ""),
+            contentType
+        );
+        requestContext.setBody(serializedBody);
+
+        let authMethod: SecurityAuthentication | undefined;
+        // Apply auth methods
+        authMethod = _config.authMethods["oauth2"]
+        if (authMethod?.applySecurityAuthentication) {
+            await authMethod?.applySecurityAuthentication(requestContext);
+        }
+        
+        const defaultAuth: SecurityAuthentication | undefined = _options?.authMethods?.default || this.configuration?.authMethods?.default
+        if (defaultAuth?.applySecurityAuthentication) {
+            await defaultAuth?.applySecurityAuthentication(requestContext);
+        }
+
+        return requestContext;
+    }
+
+    /**
+     * Get the status of an extraction by the `taskId` returned from the initial `extract/async` request.
+     * Get extraction status
+     * @param taskId The extraction task ID returned by the initial &#x60;extract/async&#x60; request.
+     */
+    public async getAsyncStatus(taskId: number, _options?: Configuration): Promise<RequestContext> {
+        let _config = _options || this.configuration;
+
+        // verify required parameter 'taskId' is not null or undefined
+        if (taskId === null || taskId === undefined) {
+            throw new RequiredError("ExtractApi", "getAsyncStatus", "taskId");
+        }
+
+
+        // Path Params
+        const localVarPath = '/cms/v3/source-code/extract/async/tasks/{taskId}/status'
+            .replace('{' + 'taskId' + '}', encodeURIComponent(String(taskId)));
+
+        // Make Request Context
+        const requestContext = _config.baseServer.makeRequestContext(localVarPath, HttpMethod.GET);
         requestContext.setHeaderParam("Accept", "application/json, */*;q=0.8")
 
 
@@ -60,13 +111,17 @@ export class ExtractApiResponseProcessor {
      * Unwraps the actual response sent by the server from the response context and deserializes the response content
      * to the expected objects
      *
-     * @params response Response returned by the server for a request to extractByPath
+     * @params response Response returned by the server for a request to doAsync
      * @throws ApiException if the response code was not in [200, 299]
      */
-     public async extractByPathWithHttpInfo(response: ResponseContext): Promise<HttpInfo<void >> {
+     public async doAsyncWithHttpInfo(response: ResponseContext): Promise<HttpInfo<TaskLocator >> {
         const contentType = ObjectSerializer.normalizeMediaType(response.headers["content-type"]);
-        if (isCodeInRange("204", response.httpStatusCode)) {
-            return new HttpInfo(response.httpStatusCode, response.headers, response.body, undefined);
+        if (isCodeInRange("202", response.httpStatusCode)) {
+            const body: TaskLocator = ObjectSerializer.deserialize(
+                ObjectSerializer.parse(await response.body.text(), contentType),
+                "TaskLocator", ""
+            ) as TaskLocator;
+            return new HttpInfo(response.httpStatusCode, response.headers, response.body, body);
         }
         if (isCodeInRange("0", response.httpStatusCode)) {
             const body: Error = ObjectSerializer.deserialize(
@@ -78,10 +133,46 @@ export class ExtractApiResponseProcessor {
 
         // Work around for missing responses in specification, e.g. for petstore.yaml
         if (response.httpStatusCode >= 200 && response.httpStatusCode <= 299) {
-            const body: void = ObjectSerializer.deserialize(
+            const body: TaskLocator = ObjectSerializer.deserialize(
                 ObjectSerializer.parse(await response.body.text(), contentType),
-                "void", ""
-            ) as void;
+                "TaskLocator", ""
+            ) as TaskLocator;
+            return new HttpInfo(response.httpStatusCode, response.headers, response.body, body);
+        }
+
+        throw new ApiException<string | Buffer | undefined>(response.httpStatusCode, "Unknown API Status Code!", await response.getBodyAsAny(), response.headers);
+    }
+
+    /**
+     * Unwraps the actual response sent by the server from the response context and deserializes the response content
+     * to the expected objects
+     *
+     * @params response Response returned by the server for a request to getAsyncStatus
+     * @throws ApiException if the response code was not in [200, 299]
+     */
+     public async getAsyncStatusWithHttpInfo(response: ResponseContext): Promise<HttpInfo<ActionResponse >> {
+        const contentType = ObjectSerializer.normalizeMediaType(response.headers["content-type"]);
+        if (isCodeInRange("200", response.httpStatusCode)) {
+            const body: ActionResponse = ObjectSerializer.deserialize(
+                ObjectSerializer.parse(await response.body.text(), contentType),
+                "ActionResponse", ""
+            ) as ActionResponse;
+            return new HttpInfo(response.httpStatusCode, response.headers, response.body, body);
+        }
+        if (isCodeInRange("0", response.httpStatusCode)) {
+            const body: Error = ObjectSerializer.deserialize(
+                ObjectSerializer.parse(await response.body.text(), contentType),
+                "Error", ""
+            ) as Error;
+            throw new ApiException<Error>(response.httpStatusCode, "An error occurred.", body, response.headers);
+        }
+
+        // Work around for missing responses in specification, e.g. for petstore.yaml
+        if (response.httpStatusCode >= 200 && response.httpStatusCode <= 299) {
+            const body: ActionResponse = ObjectSerializer.deserialize(
+                ObjectSerializer.parse(await response.body.text(), contentType),
+                "ActionResponse", ""
+            ) as ActionResponse;
             return new HttpInfo(response.httpStatusCode, response.headers, response.body, body);
         }
