+++ codegen/cms/source_code/models/AssetFileMetadata.ts
@@ -1,6 +1,6 @@
 /**
  * CMS Source Code
- * Endpoints for interacting with files in the CMS Developer File System. These files include HTML templates, CSS, JS, modules, and other assets which are used to create CMS content.
+ * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  *
  * OpenAPI spec version: v3
  * 
@@ -11,9 +11,6 @@
  */
 
 
-/**
-* The object metadata of a file.
-*/
 export class AssetFileMetadata {
     /**
     * Timestamp of when the object was first created.
@@ -52,7 +49,7 @@ export class AssetFileMetadata {
             "name": "createdAt",
             "baseName": "createdAt",
             "type": "number",
-            "format": "int32"
+            "format": "int64"
         },
         {
             "name": "archivedAt",
@@ -94,7 +91,7 @@ export class AssetFileMetadata {
             "name": "updatedAt",
             "baseName": "updatedAt",
             "type": "number",
-            "format": "int32"
+            "format": "int64"
         }    ];
 
     static getAttributeTypeMap() {
