+++ codegen/cms/source_code/models/TaskLocator.ts
@@ -1,6 +1,6 @@
 /**
  * CMS Source Code
- * Endpoints for interacting with files in the CMS Developer File System. These files include HTML templates, CSS, JS, modules, and other assets which are used to create CMS content.
+ * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  *
  * OpenAPI spec version: v3
  *
