+++ codegen/cms/source_code/types/ObservableAPI.ts
@@ -63,7 +63,7 @@ export class ObservableContentApi {
      * Create a file
      * @param environment The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
      * @param path The file system location of the file.
-     * @param file The file to upload.
+     * @param file 
      */
     public createWithHttpInfo(environment: string, path: string, file?: HttpFile, _options?: Configuration): Observable<HttpInfo<AssetFileMetadata>> {
         const requestContextPromise = this.requestFactory.create(environment, path, file, _options);
@@ -89,7 +89,7 @@ export class ObservableContentApi {
      * Create a file
      * @param environment The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
      * @param path The file system location of the file.
-     * @param file The file to upload.
+     * @param file 
      */
     public create(environment: string, path: string, file?: HttpFile, _options?: Configuration): Observable<AssetFileMetadata> {
         return this.createWithHttpInfo(environment, path, file, _options).pipe(map((apiResponse: HttpInfo<AssetFileMetadata>) => apiResponse.data));
@@ -100,7 +100,7 @@ export class ObservableContentApi {
      * Create or update a file
      * @param environment The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
      * @param path The file system location of the file.
-     * @param file The file to upload.
+     * @param file 
      */
     public createOrUpdateWithHttpInfo(environment: string, path: string, file?: HttpFile, _options?: Configuration): Observable<HttpInfo<AssetFileMetadata>> {
         const requestContextPromise = this.requestFactory.createOrUpdate(environment, path, file, _options);
@@ -126,7 +126,7 @@ export class ObservableContentApi {
      * Create or update a file
      * @param environment The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
      * @param path The file system location of the file.
-     * @param file The file to upload.
+     * @param file 
      */
     public createOrUpdate(environment: string, path: string, file?: HttpFile, _options?: Configuration): Observable<AssetFileMetadata> {
         return this.createOrUpdateWithHttpInfo(environment, path, file, _options).pipe(map((apiResponse: HttpInfo<AssetFileMetadata>) => apiResponse.data));
@@ -186,12 +186,12 @@ export class ObservableExtractApi {
     }
 
     /**
-     * Extracts a zip file in the file system. The zip file will be extracted in-place and not be deleted automatically.
-     * Extracts a zip file
-     * @param path The file system location of the zip file.
+     * Extract a zip file in the developer file system. Extraction status can be checked with the `/extract/async/tasks/taskId/status` endpoint below.
+     * Extract a zip file
+     * @param fileExtractRequest 
      */
-    public extractByPathWithHttpInfo(path: string, _options?: Configuration): Observable<HttpInfo<void>> {
-        const requestContextPromise = this.requestFactory.extractByPath(path, _options);
+    public doAsyncWithHttpInfo(fileExtractRequest: FileExtractRequest, _options?: Configuration): Observable<HttpInfo<TaskLocator>> {
+        const requestContextPromise = this.requestFactory.doAsync(fileExtractRequest, _options);
 
         // build promise chain
         let middlewarePreObservable = from<RequestContext>(requestContextPromise);
@@ -205,17 +205,50 @@ export class ObservableExtractApi {
                 for (let middleware of this.configuration.middleware) {
                     middlewarePostObservable = middlewarePostObservable.pipe(mergeMap((rsp: ResponseContext) => middleware.post(rsp)));
                 }
-                return middlewarePostObservable.pipe(map((rsp: ResponseContext) => this.responseProcessor.extractByPathWithHttpInfo(rsp)));
+                return middlewarePostObservable.pipe(map((rsp: ResponseContext) => this.responseProcessor.doAsyncWithHttpInfo(rsp)));
             }));
     }
 
     /**
-     * Extracts a zip file in the file system. The zip file will be extracted in-place and not be deleted automatically.
-     * Extracts a zip file
-     * @param path The file system location of the zip file.
+     * Extract a zip file in the developer file system. Extraction status can be checked with the `/extract/async/tasks/taskId/status` endpoint below.
+     * Extract a zip file
+     * @param fileExtractRequest 
      */
-    public extractByPath(path: string, _options?: Configuration): Observable<void> {
-        return this.extractByPathWithHttpInfo(path, _options).pipe(map((apiResponse: HttpInfo<void>) => apiResponse.data));
+    public doAsync(fileExtractRequest: FileExtractRequest, _options?: Configuration): Observable<TaskLocator> {
+        return this.doAsyncWithHttpInfo(fileExtractRequest, _options).pipe(map((apiResponse: HttpInfo<TaskLocator>) => apiResponse.data));
+    }
+
+    /**
+     * Get the status of an extraction by the `taskId` returned from the initial `extract/async` request.
+     * Get extraction status
+     * @param taskId The extraction task ID returned by the initial &#x60;extract/async&#x60; request.
+     */
+    public getAsyncStatusWithHttpInfo(taskId: number, _options?: Configuration): Observable<HttpInfo<ActionResponse>> {
+        const requestContextPromise = this.requestFactory.getAsyncStatus(taskId, _options);
+
+        // build promise chain
+        let middlewarePreObservable = from<RequestContext>(requestContextPromise);
+        for (let middleware of this.configuration.middleware) {
+            middlewarePreObservable = middlewarePreObservable.pipe(mergeMap((ctx: RequestContext) => middleware.pre(ctx)));
+        }
+
+        return middlewarePreObservable.pipe(mergeMap((ctx: RequestContext) => this.configuration.httpApi.send(ctx))).
+            pipe(mergeMap((response: ResponseContext) => {
+                let middlewarePostObservable = of(response);
+                for (let middleware of this.configuration.middleware) {
+                    middlewarePostObservable = middlewarePostObservable.pipe(mergeMap((rsp: ResponseContext) => middleware.post(rsp)));
+                }
+                return middlewarePostObservable.pipe(map((rsp: ResponseContext) => this.responseProcessor.getAsyncStatusWithHttpInfo(rsp)));
+            }));
+    }
+
+    /**
+     * Get the status of an extraction by the `taskId` returned from the initial `extract/async` request.
+     * Get extraction status
+     * @param taskId The extraction task ID returned by the initial &#x60;extract/async&#x60; request.
+     */
+    public getAsyncStatus(taskId: number, _options?: Configuration): Observable<ActionResponse> {
+        return this.getAsyncStatusWithHttpInfo(taskId, _options).pipe(map((apiResponse: HttpInfo<ActionResponse>) => apiResponse.data));
     }
 
 }
@@ -275,82 +308,6 @@ export class ObservableMetadataApi {
 
 }
 
-import { SourceCodeExtractApiRequestFactory, SourceCodeExtractApiResponseProcessor} from "../apis/SourceCodeExtractApi";
-export class ObservableSourceCodeExtractApi {
-    private requestFactory: SourceCodeExtractApiRequestFactory;
-    private responseProcessor: SourceCodeExtractApiResponseProcessor;
-    private configuration: Configuration;
-
-    public constructor(
-        configuration: Configuration,
-        requestFactory?: SourceCodeExtractApiRequestFactory,
-        responseProcessor?: SourceCodeExtractApiResponseProcessor
-    ) {
-        this.configuration = configuration;
-        this.requestFactory = requestFactory || new SourceCodeExtractApiRequestFactory(configuration);
-        this.responseProcessor = responseProcessor || new SourceCodeExtractApiResponseProcessor();
-    }
-
-    /**
-     * @param fileExtractRequest 
-     */
-    public doAsyncWithHttpInfo(fileExtractRequest: FileExtractRequest, _options?: Configuration): Observable<HttpInfo<TaskLocator>> {
-        const requestContextPromise = this.requestFactory.doAsync(fileExtractRequest, _options);
-
-        // build promise chain
-        let middlewarePreObservable = from<RequestContext>(requestContextPromise);
-        for (let middleware of this.configuration.middleware) {
-            middlewarePreObservable = middlewarePreObservable.pipe(mergeMap((ctx: RequestContext) => middleware.pre(ctx)));
-        }
-
-        return middlewarePreObservable.pipe(mergeMap((ctx: RequestContext) => this.configuration.httpApi.send(ctx))).
-            pipe(mergeMap((response: ResponseContext) => {
-                let middlewarePostObservable = of(response);
-                for (let middleware of this.configuration.middleware) {
-                    middlewarePostObservable = middlewarePostObservable.pipe(mergeMap((rsp: ResponseContext) => middleware.post(rsp)));
-                }
-                return middlewarePostObservable.pipe(map((rsp: ResponseContext) => this.responseProcessor.doAsyncWithHttpInfo(rsp)));
-            }));
-    }
-
-    /**
-     * @param fileExtractRequest 
-     */
-    public doAsync(fileExtractRequest: FileExtractRequest, _options?: Configuration): Observable<TaskLocator> {
-        return this.doAsyncWithHttpInfo(fileExtractRequest, _options).pipe(map((apiResponse: HttpInfo<TaskLocator>) => apiResponse.data));
-    }
-
-    /**
-     * @param taskId 
-     */
-    public getAsyncStatusWithHttpInfo(taskId: number, _options?: Configuration): Observable<HttpInfo<ActionResponse>> {
-        const requestContextPromise = this.requestFactory.getAsyncStatus(taskId, _options);
-
-        // build promise chain
-        let middlewarePreObservable = from<RequestContext>(requestContextPromise);
-        for (let middleware of this.configuration.middleware) {
-            middlewarePreObservable = middlewarePreObservable.pipe(mergeMap((ctx: RequestContext) => middleware.pre(ctx)));
-        }
-
-        return middlewarePreObservable.pipe(mergeMap((ctx: RequestContext) => this.configuration.httpApi.send(ctx))).
-            pipe(mergeMap((response: ResponseContext) => {
-                let middlewarePostObservable = of(response);
-                for (let middleware of this.configuration.middleware) {
-                    middlewarePostObservable = middlewarePostObservable.pipe(mergeMap((rsp: ResponseContext) => middleware.post(rsp)));
-                }
-                return middlewarePostObservable.pipe(map((rsp: ResponseContext) => this.responseProcessor.getAsyncStatusWithHttpInfo(rsp)));
-            }));
-    }
-
-    /**
-     * @param taskId 
-     */
-    public getAsyncStatus(taskId: number, _options?: Configuration): Observable<ActionResponse> {
-        return this.getAsyncStatusWithHttpInfo(taskId, _options).pipe(map((apiResponse: HttpInfo<ActionResponse>) => apiResponse.data));
-    }
-
-}
-
 import { ValidationApiRequestFactory, ValidationApiResponseProcessor} from "../apis/ValidationApi";
 export class ObservableValidationApi {
     private requestFactory: ValidationApiRequestFactory;
@@ -371,7 +328,7 @@ export class ObservableValidationApi {
      * Validates the file contents passed to the endpoint given a specified path and environment. Accepts multipart/form-data content type.
      * Validate the contents of a file
      * @param path The file system location of the file.
-     * @param file The file to validate.
+     * @param file 
      */
     public doValidateWithHttpInfo(path: string, file?: HttpFile, _options?: Configuration): Observable<HttpInfo<void>> {
         const requestContextPromise = this.requestFactory.doValidate(path, file, _options);
@@ -396,7 +353,7 @@ export class ObservableValidationApi {
      * Validates the file contents passed to the endpoint given a specified path and environment. Accepts multipart/form-data content type.
      * Validate the contents of a file
      * @param path The file system location of the file.
-     * @param file The file to validate.
+     * @param file 
      */
     public doValidate(path: string, file?: HttpFile, _options?: Configuration): Observable<void> {
         return this.doValidateWithHttpInfo(path, file, _options).pipe(map((apiResponse: HttpInfo<void>) => apiResponse.data));
